Fix/json payload fix (#59)
Fixed json payload for Windows, linux and MAc. Fixed doc generation for Windows and Linux. However, this solution is currently not supported on MacOS. This is due to the lack of support for 'AWT' in GraalVM Native Images on MacOS, as explained in this GitHub issue: oracle/graal#4124. We are working on providing support for MacOS in the future.